MSVC porting work (in progress).
* redirect GCC __attribute__((unused))
* redirect <sys/cdefs.h>
* redirect <sys/param.h>
* replace (some instances of) bzero, bcopy
* block <err.h> when _MSC_VER
* fix returns from strlen in xdr_inline.h and xdr.c
* implement atomic_{add,sub}_uint32_t using intrinsics
* replace C99 "static inline" with macro __static_inline, and define
* include WSA headers in <misc/portable.h>
* conditionally hook MSVC CRT debugging
* type fixes in winpthreads.h
* redirect poll and ioctl

+++ src/xdr.c
@@ -657,7 +657,7 @@ xdr_string(XDR *xdrs, char **cpp, u_int maxsize)
     case XDR_ENCODE:
         if (sp == NULL)
             return FALSE;
-        size = strlen(sp);
+        size = (u_int) strlen(sp);
         break;
     case XDR_DECODE:
         break;
